+++ src/app/(auth)/mypage/achievement/page.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 import clsx from 'clsx';
-import { getUserProfile, getUserSessionState } from '@/lib/utils/api/auth-action';
+import { getUserProfile } from '@/lib/utils/api/auth-action';
 import { getUserMission } from '@/lib/utils/api/my-mission.api';
 import { calculateUserLevel } from '@/lib/utils/calculate-user-level';
 import { getLevelsByTypes } from '@/lib/utils/get-level-by-types';
@@ -12,6 +12,7 @@ import TRAVEL_ICON from '@images/images/travel-icon.svg';
 import GOAT_ICON from '@images/images/goat-icon.svg';
 import CLEAN_ICON from '@images/images/clean-icon.svg';
 import PLAY_ICON from '@images/images/play-icon.svg';
+import { getUserSessionState } from '@/lib/utils/api/auth.api';
 
 const AchievementPage = async () => {
   //유저 닉네임 및 로그인 조회
