+++ src/lib/utils/api/my-mission.api.ts
@@ -1,7 +1,7 @@
 import { TABLE } from '@/constants/supabase-tables-name';
 import { MSG } from '@/constants/messages';
 import { createClient } from '@/lib/utils/supabase/supabase-server';
-import { getUserSessionState } from '@/lib/utils/api/auth-action';
+import { getUserSessionState } from '@/lib/utils/api/auth.api';
 
 /**
  * Supabase에서 `user_mission` 테이블의 `로그인 한 User`가 달성한 미션 전부를 조회하는 함수
